feat : Classe css pour confirmer une suppression
Ajoute la classe css "modal-delete-confirm" qui permettra dans le futur
d'afficher facilement une fenêtre de confirmation pour la suppression

+++ src/Spring/webApp/src/main/resources/templates/view/pays/lister.ftl
@@ -18,7 +18,7 @@
 				<tr>
 					<td>${p.libelle}</td>
 					<td class="rigth table-actions">
-						<a href="/pays/${p.id}/update"><i class="material-icons">create</i></a><a href="/pays/${p.id}/delete"><i class="material-icons">delete</i></a>
+						<a href="/pays/${p.id}/update"><i class="material-icons">create</i></a><a href="/pays/${p.id}/delete" class="modal-delete-confirm"><i class="material-icons">delete</i></a>
 					</td>
 				</tr>
 			</#list>
